- Added authorization for users to access reviews (Need to reconsider REST API path design)

+++ src/main/java/dev/daniel/BooksGalore/Service/ReviewService.java
@@ -1,6 +1,9 @@
 package dev.daniel.BooksGalore.Service;
 
+import dev.daniel.BooksGalore.Exception.ReviewNotFoundException;
+import dev.daniel.BooksGalore.Exception.UnauthorizedException;
 import dev.daniel.BooksGalore.Model.Review;
+import dev.daniel.BooksGalore.Model.User;
 import dev.daniel.BooksGalore.Repository.ReviewRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,14 +19,25 @@ public class ReviewService {
         return repo.findAllByBook_ISBNid(ISBNid);
     }
 
-    public Review addReview(Review review){
+    public Review addReview(Review review, User authenticatedUser){
+        review.setUser(authenticatedUser);
         return repo.save(review);
     }
 
-    public Review updateReview(Review review){
+    public Review updateReview(int id, Review review ,User authenticatedUser){
+        Review existingReview = repo.findById(id).orElseThrow(() -> new ReviewNotFoundException("Review not found"));
+        if(!existingReview.getUser().equals(authenticatedUser)){
+            throw new UnauthorizedException("You are not authorized to edit this review");
+        }
+        review.setId(id);
+        review.setUser(authenticatedUser);
         return repo.save(review);
     }
-    public void deleteReview(int id){
+    public void deleteReview(int id, User authenticatedUser){
+        Review review = repo.findById(id).orElseThrow(() -> new ReviewNotFoundException("Review not found"));
+        if(!review.getUser().equals(authenticatedUser)){
+            throw new UnauthorizedException("You are not authorized to delete this review");
+        }
         repo.deleteById(id);
     }
